+++ src/NugetNinja.DeprecatedPackagePlugin/NugetNinja.DeprecatedPackagePlugin.csproj
@@ -8,10 +8,6 @@
     <RootNamespace>Microsoft.NugetNinja.PossiblePackageUpgradePlugin</RootNamespace>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="6.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" />
-  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\NugetNinja.Core\NugetNinja.Core.csproj" />
