+++ src/NugetNinja.UselessPackageReferencePlugin/StartUp.cs
@@ -3,7 +3,6 @@
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.NugetNinja.Core;
-using Microsoft.NugetNinja.Core;
 
 namespace Microsoft.NugetNinja.UselessPackageReferencePlugin;
 
@@ -11,8 +10,6 @@ public class StartUp : IStartUp
 {
     public void ConfigureServices(IServiceCollection services)
     {
-        services.AddTransient<Extractor>();
-        services.AddTransient<ProjectsEnumerator>();
         services.AddTransient<UselessPackageReferenceDetector>();
     }
 }
